Rename port to tcp and not export it through the cli-kit's index.ts
