+++ packages/app/src/cli/services/dev.ts
@@ -12,10 +12,11 @@ import {AppInterface, AppConfiguration, Web, WebType} from '../models/app/app.js
 import metadata from '../metadata.js'
 import {UIExtension} from '../models/app/extensions.js'
 import {fetchProductVariant} from '../utilities/extensions/fetch-product-variant.js'
-import {analytics, output, port, system, session, abort, string, environment} from '@shopify/cli-kit'
+import {analytics, output, system, session, abort, string, environment} from '@shopify/cli-kit'
 import {Config} from '@oclif/core'
 import {execCLI2} from '@shopify/cli-kit/node/ruby'
 import {renderConcurrent} from '@shopify/cli-kit/node/ui'
+import {getAvailableTCPPort} from '@shopify/cli-kit/node/tcp'
 import {Writable} from 'node:stream'
 
 export interface DevOptions {
@@ -67,7 +68,7 @@ async function dev(options: DevOptions) {
     cachedTunnelPlugin: tunnelPlugin,
   })
 
-  const backendPort = await port.getRandomPort()
+  const backendPort = await getAvailableTCPPort()
 
   const frontendConfig = options.app.webs.find(({configuration}) => configuration.type === WebType.Frontend)
   const backendConfig = options.app.webs.find(({configuration}) => configuration.type === WebType.Backend)
@@ -101,7 +102,7 @@ async function dev(options: DevOptions) {
   }
 
   const proxyTargets: ReverseHTTPProxyTarget[] = []
-  const proxyPort = usingLocalhost ? await port.getRandomPort() : frontendPort
+  const proxyPort = usingLocalhost ? await getAvailableTCPPort() : frontendPort
   const proxyUrl = usingLocalhost ? `${frontendUrl}:${proxyPort}` : frontendUrl
 
   if (options.app.extensions.ui.length > 0) {
