+++ packages/app/src/cli/services/dev/urls.test.ts
@@ -12,8 +12,11 @@ import {api, environment, error, outputMocker, plugins, store, ui} from '@shopif
 import {Config} from '@oclif/core'
 import {err, ok} from '@shopify/cli-kit/common/result'
 import {AbortSilentError, BugError} from '@shopify/cli-kit/node/error'
+import {getAvailableTCPPort} from '@shopify/cli-kit/node/tcp'
 
 beforeEach(() => {
+  vi.mock('@shopify/cli-kit/node/tcp')
+  vi.mocked(getAvailableTCPPort).mockResolvedValue(3042)
   vi.mock('@shopify/cli-kit', async () => {
     const cliKit: any = await vi.importActual('@shopify/cli-kit')
     return {
@@ -36,9 +39,6 @@ beforeEach(() => {
       },
       store: {
         setAppInfo: vi.fn(),
-      },
-      port: {
-        getRandomPort: async () => 3042,
       },
       environment: {
         local: {
