+++ packages/app/src/cli/services/dev/urls.ts
@@ -1,8 +1,9 @@
 import {updateURLsPrompt} from '../../prompts/dev.js'
 import {AppInterface} from '../../models/app/app.js'
-import {api, environment, output, plugins, port, store} from '@shopify/cli-kit'
+import {api, environment, output, plugins, store} from '@shopify/cli-kit'
 import {AbortError, AbortSilentError, BugError} from '@shopify/cli-kit/node/error'
 import {Config} from '@oclif/core'
+import {getAvailableTCPPort} from '@shopify/cli-kit/node/tcp'
 
 export interface PartnersURLs {
   applicationUrl: string
@@ -75,10 +76,10 @@ export async function generateFrontendURL(options: FrontendURLOptions): Promise<
   }
 
   if (needsTunnel) {
-    frontendPort = await port.getRandomPort()
+    frontendPort = await getAvailableTCPPort()
     frontendUrl = await generateURL(options.commandConfig, frontendPort)
   } else {
-    frontendPort = await port.getRandomPort()
+    frontendPort = await getAvailableTCPPort()
     frontendUrl = 'http://localhost'
     usingLocalhost = true
   }
