+++ packages/cli-kit/src/index.ts
@@ -16,7 +16,6 @@ export * as os from './os.js'
 export * as output from './output.js'
 export * as path from './path.js'
 export * as plugins from './plugins.js'
-export * as port from './port.js'
 export * as schema from './schema.js'
 export * as session from './session.js'
 export * as store from './store.js'
